feat: apply clickable filename format to codeframeFormatter
When an error message is shown on an IDE console, most IDEs will allow
you to ctrl+click the file name to go directly to the proper file, line
and character. In order for this to work properly, the error filename,
line and character must have a specific format: "filename.js(line,col)".
The codeFormatter already follows this format, this commit applies this
format to the codeframeFormatter.