feat: add issue.scope option (#442)
The plugin uses TypeScript settings regarding files that are included in the compilation. It can be a different set of files comparing to the webpack compilation. To not display issues that are related to files outside webpack compilation, an `issue.scope` option has been added.

BREAKING CHANGE: 🧨 Issues outside webpack compilation will not be reported by default. See `issue.scope` option.