feat: remove "incremental: true" from ts default options
As incremental mode has some assumptions regarding changes that can
occur in the project, sometimes the user has to remove .tsbuildinfo
files manually to fix it. Because of that, we should not enable
incremental mode by default - it should be a conscious choice of the
user.

BREAKING CHANGE: 🧨 Remove "incremental: true" from default typescript options