fix: use memfs for ts artifacts to prevent mtime conflict (#617)
When ts-loader uses projectReferences mode, it writes down .js and .d.ts
files. If it does it faster than the plugin, the typescript instance in
the plugin is confused and doesn't work properly. To fix that, let's use
memfs for dirs and files that are detected as a project artifacts.