feat: drop support for Vue.js
This commit drops support for Vue.js in the plugin. For the last few
years, support for Vue.js has been broken, and the implementation of
Vue.js support was a hacky work-around. Unfortunately, TypeScript
doesn't expose an API to do that properly, and while I created a feature
proposal API on the TypeScript repository, the TypeScript team has other
priorities. As I won't have time to support this feature and am
receiving justified bug reports, I believe it's better to make it clear
that Vue.js is not supported by the plugin instead of pretending that it
is.

BREAKING CHANGE: 🧨 Drop support for Vue.js